moved util functions from type file
