+++ src/Ens.ts
@@ -1,7 +1,7 @@
 import { default as ensInterface } from './ens/contract/ens';
 import { default as resolverInterface } from './ens/contract/resolver';
 import { formatsByCoinType } from '@ensdomains/address-encoder';
-import { EthCoinIndex, Bip44Constants, isNullAddress } from './types';
+import { EthCoinIndex, Bip44Constants } from './types';
 import { EthereumNamingService } from './EthereumNamingService';
 import {
   NamingServiceName,
@@ -12,6 +12,7 @@ import Contract from './utils/contract';
 import contentHash from 'content-hash';
 import EnsNetworkMap from 'ethereum-ens-network-map';
 import { ResolutionResponse, CryptoRecords, SourceDefinition } from './publicTypes';
+import { isNullAddress } from './utils';
 
 export default class Ens extends EthereumNamingService {
   readonly name = NamingServiceName.ENS;
